requirements.txt
asgiref==3.9.1
Django==5.2.6
django-rest-framework==0.1.0
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
